Fix message creator not re-rendering when the chat changes
The message creator didn't re-render when the user switched
between chats because the <useState()> hook retained the
previous chat's state since React didn't know that the page
had actually changed. Even though we're supposed to use
React keys only in the context of an array or sibling
elements, it was the only practical solution in this case
because otherwise we'd have to have complex logic in each
of the child, sub-child, etc. elements regarding state
updates (which is more of a bug than using React keys
"incorrectly").